[native] Add a native implementation of the tokenizer
Ports CPython's tokenize.c to rust, and adds features to it to (mostly)
match the behavior of the Parso-based tokenizer.

Some tests are currently failing, but most of this is just due to
differences between error handing in the Python in Rust implementations.
I'll either fix or disable these cases in the future.

Like the whitespace_parser commit, this appears to make parsing faster
by a small margin, but tokenizing is just one small part of the overall
parsing process.